Refactor to only get usage data
